build(renderFunc): create a render function to render all shapes in index.js, move the shapFactory function to index.js
