+++ lib/childShapes.js
@@ -85,23 +85,4 @@ class Square extends Shape {
 
 };
 
-const shapeFactory = (type, text, textColor, shapeColor, bgColor, ...args) => {
-
-  switch (type) { 
-    case "triangle":
-      return new Triangle(text, textColor, shapeColor, bgColor, ...args);
-    case "circle":
-      return new Circle(text, textColor, shapeColor, bgColor, ...args);
-    case "square":
-      return new Square(text, textColor, shapeColor, bgColor, ...args);
-    default:
-      throw new TypeError("Please enter a valid type");
-  };
-
-};
-
-Triangle.render();
-Circle.render();
-Square.render();
-
-module.exports = shapeFactory;
+module.exports = { Triangle, Circle, Square }; // Exporting the Triangle, Circle, and Square classes
